refactor: create header global component
